requirements.txt
flask
request
gunicorn
dialogflow
twilio
gnewsclient
